color dump and rethinking the args handling
I created two wrappers image_load and image_free that implement
the load and dealocating features of the library stb_image.
The type Image will store anything related to the loaded image.

I also encapsulated the cmd input flags in a struct Args to
make it clearer when we are dealing with command line options.

Implemented the dump colors with the option to dump them in
hexadecimal.

I still need to modularize it, my idea is creating the files
src/image/image.c and src/image/image.h to handle everything
related to images and stb and the core modules should access
only a simple interface.