Fixed segmentation fault on dealocation
when the user passed -d without informing the filename the program
would try to dealocate the inexistent pointer img_p->data causing
a segmentation fault.

Added some checks and improved the way to allocate so we only
have to declare a pointer to Image in the main funcion with no
risk of non initialized values.