Creating auxiliar functions to handle image data
I created the data type Pixel that can store a pixel
up to 4 channels

it stores [r, g, b, a] if rgba
          [r, g, b, 0] if rgb
          [v, 0, 0, 0] if grayscale with one channel
          [v, a, 0, 0] if grayscale with transparency

I created the function image_get_pixel that receives
the pointer to the image, the point (x,y) and the
pointer to the Pixel object where the data will be
stored.

I also created the functions image_pixel_to_32 and
image_32_to_pixel that convert the pixel type to
a 32 bits number. This might be useful to display the
hexadecimal value of the color and to use the
color as a key in some hash function.