Creating functions to manipulate a string array
I want the dump colors function to print all unique colors
and I will handle that with a hashmap. The one at search.h.

For that, I need to keep track of a array of string keys.
This commit provides the necessary tools to manipulate an
array of strings.

I created a new module src/utils that will have all the
data structures and annoying code that are not related to
the business logic.

Inside that, I created the module arrstr.h, that has three
functions:

    arr_str_push        add a string to an array char**
    arr_str_delete      delete position from an array char**
    arr_str_free        dealocate memory of an array char**

More details in the comments of arrstr.h

I also added some rubish comments of me trying to learn how to
use the hsearch function. Sorry about that.