Created some basic functions for kmeans
I created some functions that are going to be used in the kmeans
algorithm. They are in the module kmeans.h.

Kpoint is a data structure that indicates the point processed
by the kmeans algorithm. The value should have an arbitrary
dimension indicated by dim, but for now it has a fixed value of 3.

I implemented the kmeans++ algorithm which is a cluster
initialization method that provides a high probability of choosing
initial clusters that are distant from each other, which is
something that will reduce a lot the iterations needed by the
kmeans.

More information about this method can be found in the article
"kmeans++: The advantages of Careful Seeding"
by David Arthur and Sergei Vassilvitskii [1].

in the main function I wrote some test code that executes when
you give the -c argument. This will be removed in future commits.

[1] http://ilpubs.stanford.edu:8090/778/1/2006-13.pdf