Created type kmeans_t and data allocation
The KMEANS_T type is a data structure where I aggregate
correlated data used in the kmeans procedure so the final API
user won't need to worry about many parameters or complex
data allocation.

It consists in:
    - dim                the dimension of the data
    - num_clusters       number of clusters
    - num_points         number of points
    - points             the array of points
    - clusters           the array of clusters

The allocation is done with the function kmeans_alloc, the
user needs to send the number of points, the number of clusters
and dimension as parameters. This function returns a pointer to
the type KMEANS_T which can be deallocated with the function
kmeans_destroy.

I still need to adapt the other functions to use this new data
type.