Now we initialize the kmeans with a double**
we now give a double** to the function kmeans_apply which is
now easier to call because of the KMEANS_T type.  We just need
to pass the initialized KMEANS_T* and the initialization method.

I also fixed some pointers issues like I was copying the reference
to points to the new centroids and this was making the calculations
go wrong.